Changes to errors and testing

+++ ft_atoi.c
@@ -12,8 +12,8 @@
 
 int	ft_atoi(const char *str)
 {
-	int			i;
-	int			plusminus;
+	long int	i;
+	int		plusminus;
 	long int	result;
 
 	i = 0;
